chore(public): update image assets and remove site webmanifest
This commit updates several binary image files in the public directory, including icons for Android Chrome and Apple touch, as well as favicons. Additionally, the site.webmanifest file has been removed, which contained metadata about the site's name, icons, theme color, and display mode. Removing the manifest may affect how the site is displayed when added to home screens or used in standalone mode on devices. Ensure that the removal of this file aligns with the current needs and functionality of the project.